Split field tests
